refactor: delete extra files
