+++ .eslintrc
@@ -8,16 +8,13 @@
   "rules": {
     "strict": 0,
     "import/prefer-default-export": "off",
-
     "react/prop-types": "off",
     "react/jsx-filename-extension": "off",
-    "react/destructuring-assignment": "off",
     "react/button-has-type": "off",
-    "react/jsx-props-no-spreading": "off",
     "react/state-in-constructor": "off",
-    "react/static-property-placement": "off",
-
-    "jsx-a11y/click-events-have-key-events": "off",
-    "jsx-a11y/no-noninteractive-element-interactions": "off"
+    // "react/static-property-placement": "off",
+    "jsx-a11y/no-static-element-interactions": "off",
+    "jsx-a11y/click-events-have-key-events": "off"
+    // "jsx-a11y/no-noninteractive-element-interactions": "off"
   }
 }
