+++ package.json
@@ -39,7 +39,6 @@
     "react-hot-loader": "^4.12.21",
     "react-player": "^2.6.0",
     "regenerator-runtime": "^0.13.7",
-    "typescript": "^3.9.7",
     "webpack": "^4.43.0",
     "webpack-dev-middleware": "^3.7.2"
   },
@@ -71,6 +70,8 @@
     "prettier": "^2.0.5",
     "sass-loader": "^9.0.2",
     "style-loader": "^1.2.1",
+    "ts-loader": "^8.0.2",
+    "typescript": "^3.9.7",
     "url-loader": "^4.1.0",
     "webpack-bundle-analyzer": "^3.8.0",
     "webpack-cli": "^3.3.12",
