+++ public/index.html
@@ -4,6 +4,8 @@
     <meta charset="utf-8" />
     <title>Songbird</title>
     <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <meta http-equiv="X-UA-Compatible" content="ie=edge" />
+    <meta http-equiv="Content-Security-Policy" content="upgrade-insecure-requests" />
     <link rel="icon" href="favicon.ico" />
     <link rel="stylesheet" href="https://bootswatch.com/4/darkly/bootstrap.min.css" />
   </head>
