--- src/components/bird-service-context/bird-service-context.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const { Provider: BirdServiceProvider, Consumer: BirdServiceConsumer } = React.createContext();
-
-export { BirdServiceProvider, BirdServiceConsumer };
