--- src/components/bird-service-context/index.js
@@ -1,3 +1,0 @@
-import { BirdServiceProvider, BirdServiceConsumer } from './bird-service-context';
-
-export { BirdServiceProvider, BirdServiceConsumer };
