+++ src/components/error-boundry/error-boundry.jsx
@@ -3,23 +3,24 @@ import React, { Component } from 'react';
 import ErrorIndicator from '../error-indicator/error-indicator';
 
 export default class ErrorBoundry extends Component {
-
   state = {
-    hasError: false
+    hasError: false,
   };
 
   componentDidCatch() {
     this.setState({
-      hasError: true
+      hasError: true,
     });
   }
 
   render() {
+    const { hasError } = this.state;
+    const { children } = this.props;
 
-    if (this.state.hasError) {
-      return <ErrorIndicator />
+    if (hasError) {
+      return <ErrorIndicator />;
     }
 
-    return this.props.children;
+    return children;
   }
 }
