+++ src/components/error-indicator/error-indicator.jsx
@@ -6,14 +6,10 @@ import icon from './death-star.png';
 const ErrorIndicator = () => {
   return (
     <div className="error-indicator">
-      <img src={icon} alt="error icon"/>
+      <img src={icon} alt="error icon" />
       <span className="boom">BOOM!</span>
-      <span>
-        something has gone terribly wrong
-      </span>
-      <span>
-        (but we already sent droids to fix it)
-      </span>
+      <span>something has gone terribly wrong</span>
+      <span>(but we already sent droids to fix it)</span>
     </div>
   );
 };
